Split apollo.rs into smaller modules.
